+++ aries-core/src/js/components/layouts/index.js
@@ -1,3 +1,4 @@
+export * from './ActionMenu';
 export * from './ButtonGroup';
 export * from './Layer';
 export * from './ModalDialog';
