+++ aries-site/src/examples/components/pageheader/PageHeaderResponsiveActions.js
@@ -9,10 +9,11 @@ import {
   ResponsiveContext,
   ThemeContext,
 } from 'grommet';
+import { ActionMenu } from 'aries-core';
 import { groupActions } from './utils';
-import { ActionMenu, ReverseAnchor } from '../../templates';
+import { ReverseAnchor } from '../../templates';
 
-// need to overrid the PageHeader theme values in order for our examples 
+// need to overrid the PageHeader theme values in order for our examples
 // to fit their container page. Even though it is using the correct breakpoints
 // our examples are only taking up half the space on the page so layout
 // different than a PageHeader on a full page.
