+++ aries-site/src/examples/templates/feedback/ModalTaskFlowExample/PowerOnFlow.js
@@ -1,8 +1,7 @@
 import { useState } from 'react';
 import { Box, Button, Spinner, Text } from 'grommet';
 import { StatusGoodSmall } from 'grommet-icons';
-import { ModalBody, ModalFooter } from 'aries-core';
-import { ButtonGroup } from 'aries-core/src/js/components/layouts/ButtonGroup';
+import { ButtonGroup, ModalBody, ModalFooter } from 'aries-core';
 
 const steps = {
   'confirm-power-on': {
