+++ aries-site/src/examples/templates/index.js
@@ -17,7 +17,6 @@ export * from './table-customization';
 export * from './toast-notifications';
 export * from './wizard';
 /* Reusable templates */
-export * from './ActionMenu';
 export * from './DestructiveConfirmation';
 export * from './FilterControls';
 export * from './ReverseAnchor';
