+++ aries-site/src/pages/components/pageheader.mdx
@@ -17,7 +17,7 @@ The PageHeader's title communicates the context of the page. The subtitle elabor
   code={[
     "https://raw.githubusercontent.com/grommet/hpe-design-system/master/aries-site/src/examples/components/pageheader/PageHeaderIntroExample.js",
     'https://raw.githubusercontent.com/grommet/hpe-design-system/master/aries-site/src/examples/templates/ReverseAnchor/ReverseAnchor.js',
-    'https://raw.githubusercontent.com/grommet/hpe-design-system/master/aries-site/src/examples/templates/ActionMenu/ActionMenu.js',
+    'https://raw.githubusercontent.com/grommet/hpe-design-system/master/aries-core/src/js/layouts/ActionMenu/ActionMenu.js',
   ]}
   docs="https://v2.grommet.io/pageheader?theme=hpe"
   figma="https://www.figma.com/file/BHp0KrOcPvbnzNyshfHkd8/HPE-Page-Header?node-id=2%3A5"
