+++ aries-site/src/pages/templates/forms/managing-child-objects.mdx
@@ -66,7 +66,7 @@ heading, optional description, and a secondary button labelled with an add actio
     'https://raw.githubusercontent.com/grommet/hpe-design-system/master/aries-site/src/examples/templates/FormChildObject/FormChildObjects.js',
     'https://raw.githubusercontent.com/grommet/hpe-design-system/master/aries-site/src/examples/templates/FormChildObject/FormChildObject.js',
     'https://raw.githubusercontent.com/grommet/hpe-design-system/master/aries-site/src/examples/templates/FormChildObject/ChildHeader.js',
-    'https://raw.githubusercontent.com/grommet/hpe-design-system/master/aries-site/src/examples/templates/ButtonGroup/ButtonGroup.js',
+    'https://raw.githubusercontent.com/grommet/hpe-design-system/master/aries-core/src/js/layouts/ButtonGroup/ButtonGroup.js',
   ]}
   github="https://github.com/grommet/hpe-design-system/blob/master/aries-site/src/examples/templates/forms/managing-child-objects/CreateRole.js"
   showResponsiveControls={[]}
@@ -85,7 +85,7 @@ child (e.g. creating a cluster) may present the first child object in its expand
     'https://raw.githubusercontent.com/grommet/hpe-design-system/master/aries-site/src/examples/templates/FormChildObject/FormChildObjects.js',
     'https://raw.githubusercontent.com/grommet/hpe-design-system/master/aries-site/src/examples/templates/FormChildObject/FormChildObject.js',
     'https://raw.githubusercontent.com/grommet/hpe-design-system/master/aries-site/src/examples/templates/FormChildObject/ChildHeader.js',
-    'https://raw.githubusercontent.com/grommet/hpe-design-system/master/aries-site/src/examples/templates/ButtonGroup/ButtonGroup.js',
+    'https://raw.githubusercontent.com/grommet/hpe-design-system/master/aries-core/src/js/layouts/ButtonGroup/ButtonGroup.js',
   ]}
   github="https://github.com/grommet/hpe-design-system/blob/master/aries-site/src/examples/templates/forms/managing-child-objects/CreateCluster.js"
   showResponsiveControls={[]}
@@ -104,7 +104,7 @@ to allow a user the ability to easily remove all associations with a single acti
     'https://raw.githubusercontent.com/grommet/hpe-design-system/master/aries-site/src/examples/templates/FormChildObject/FormChildObjects.js',
     'https://raw.githubusercontent.com/grommet/hpe-design-system/master/aries-site/src/examples/templates/FormChildObject/FormChildObject.js',
     'https://raw.githubusercontent.com/grommet/hpe-design-system/master/aries-site/src/examples/templates/FormChildObject/ChildHeader.js',
-    'https://raw.githubusercontent.com/grommet/hpe-design-system/master/aries-site/src/examples/templates/ButtonGroup/ButtonGroup.js',
+    'https://raw.githubusercontent.com/grommet/hpe-design-system/master/aries-core/src/js/layouts/ButtonGroup/ButtonGroup.js',
   ]}
   github="https://github.com/grommet/hpe-design-system/blob/master/aries-site/src/examples/templates/forms/managing-child-objects/EditRole.js"
   showResponsiveControls={[]}
