+++ aries-site/src/pages/templates/user-feedback-collection.mdx
@@ -155,7 +155,7 @@ power on task flow.
     'https://raw.githubusercontent.com/grommet/hpe-design-system/master/aries-site/src/examples/templates/feedback/ModalTaskFlowExample/PowerOnFlow.js',
     'https://raw.githubusercontent.com/grommet/hpe-design-system/master/aries-site/src/examples/templates/feedback/ModalTaskFlowExample/TaskFlowFeedback.js',
     'https://raw.githubusercontent.com/grommet/hpe-design-system/master/aries-core/src/js/layouts/ModalDialog/ModalDialog.js',
-    'https://raw.githubusercontent.com/grommet/hpe-design-system/master/aries-site/src/examples/templates/ButtonGroup/ButtonGroup.js',
+    'https://raw.githubusercontent.com/grommet/hpe-design-system/master/aries-core/src/js/layouts/ButtonGroup/ButtonGroup.js',
   ]}
   github="https://raw.githubusercontent.com/grommet/hpe-design-system/master/aries-site/src/examples/templates/feedback/ModalTaskFlowExample/ModalTaskFlowExample.js"
   showResponsiveControls={['fullScreen']}
